requirements.txt
fastapi
uvicorn
vosk
python-dotenv
